refactor(evaluations): extraire EvaluationCard de TeacherEvaluations (#28)
Tranche 3/3 de la décomposition de TeacherEvaluations.vue.

- EvaluationCard.vue (642 l.) : sous-composant de présentation d'une carte
  d'évaluation. Props evaluation/syncing, émet create/edit/view-results/
  publish/preview/sync/delete. Embarque getStatusIcon + formatDate + sa CSS
  scopée (copiée à l'identique).
- La vue utilise <EvaluationCard v-for> ; markup + CSS carte + helpers +
  icônes inutilisées retirés. 1654 -> 1065 l.

Cumul décomposition : 1877 -> 1065 l. (-43 %). Script entièrement décomposé
(logique pure #44, composable #45, carte ici).

Dette tracée : EvaluationCard 642 l. (surtout CSS) dépasse encore §1.1 ;
modale/stats/filtres restent inline (levier moindre). Parité visuelle : CSS
déplacée à l'identique + build vert, QA manuelle recommandée (pas de test UI).

Tests 180/180, contract 28/28, build OK.
